Admin Panel Beginnings / Symbol Uploader
And a few other things, FineUploader has been included in the static lib
right now because apparently you can't just have it prebuilt from bower
so I'm saving people the pain in the ass that is doing it yourself.

Also deleted a bunch of old admin route files as they're not needed
anymore with the AngularJS setup, the admin panel is by no means done
but it does work for the time being, also there is now a commandline
argument 'enable_create_admin' that allows you to set a secret key to
create new administrator accounts with POST requests, obviously this
shouldn't be enabled after initial setup. (It is created at
/admin/create_admin/:option and requires post arguments 'username'
'password' and 'email'